requirements.txt
pytest==6.1.2
pyyaml==5.3.1
ttkthemes==3.1.1
